+++ alteia/apis/client/auth/sharetokensimpl.py
@@ -63,8 +63,7 @@ class ShareTokensImpl:
         if len(company_ids) != 1:
             raise RuntimeError('Expecting datasets in a single company')
 
-        dataset_obj = self._sdk.datasets.describe(dataset)
-        data['company'] = dataset_obj.company
+        data['company'] = list(company_ids)[0]
 
         if duration is not None:
             data['duration'] = duration